[FEATURE] Java-driver-4 | Improved concurrency and peer enumeration in CosmosLoadBalancingPolicy
Concurrency improvements

We used to rely on CopyOnWriteArrayList. The implementation locked and copied often while not ensuring that the lists that should be updated concurrently were updated concurrently. The new implementation corrects this using ArrayList and locks explicitly.

Peer enumeration improvements

The Datastax Java Driver does not guarantee that peers will be enumerated prior to the first read/write operation and--until this PR--was likely to cause a NoNodeAvailableException when querying a multi-region Cosmos DB Cassandra instance. This PR Updates CosmosLoadBalancingPolicy::init and CosmosLoadBalancingPolicy::newQueryPlan to guarantee that the list of peers is fully enumerated before CosmosLoadBalancingPolicy::newQueryPlan executes for the first time. This ensures that all regions are enumerated and then considered based on their distance--whether local or remote--avoiding these NoNodeAvailableException failures.

The fix in this PR is to refresh all nodes on the call to CosmosLoadBalancingPolicy::init and--until this short-running operation is complete--queue subsequent calls to CosmosLoadBalancingPolicy::newQueryPlan. This is done using a semaphore with a lifetime that ends when the operation is complete.

This is different than what's done by DefaultLoadBalancingPolicy. It refreshes based on routing requirements that are different than we have. It relies on the token map which we disable because token maps aren't relevant to our routing.

See https://github.com/Azure/azure-cosmos-cassandra-extensions/pull/16 for a list of additional changes.