CosmosLoadBalancingPolicy based on preferreed-regions and a determination that multi-region-writes are enabled.
Today we let the user tell us with a setting. In the near future we'll add a metadata to the cosmos-specific system table to query for the setting.

This is the same strategy used by the .NET SDK. It produces predictable failover behavior and is easily understood. No settings are required. We do failover in alpha order
when no preferred-region list is provided.

This implementation also covers usage when--by chance--the user happens to connect to an Apache Cassandra instance. We honor the preferred region list and do not mind
that a region may have more than one contact-point. We'll work, but--I am assuming--not as well as the default load-balancing-policy. This is no different than our previous
implementation. Neither implements token-based routing.