CosmosLoadBalancingPolicy based on preferred regions that define the order in which failover occurs (#26)
This implementation is the same strategy as used by the Java and .NET SDKs for Cosmos DB SQL. We ask the user to specify a list of preferred-regions and set a flag indicating whether multi-region-writes are enabled. In the near future I propose we add metadata to the cosmos-specific system table that will indicate other multi region writes are enabled.

This strategy produces predictable failover behavior and is easily understood. No settings are required. We do failover in alpha order when no preferred-region list is provided. We always fail over to the global endpoint.

This implementation also covers usage when--by chance--the user happens to connect to an Apache Cassandra instance. We honor the preferred region list and do not mind that a region may have more than one contact-point. We'll work, but--I am assuming--not as well as the default load-balancing-policy when used on an Apache Cassandra instance. This is no different than our previous implementation. Neither implements token-based routing or uses metrics to identify slow and fast nodes, avoiding the former and preferring the latter.