[FIX] Addresses two CosmosLoadBalancingPolicy issues and Log4j2 Vulnerability “Log4Shell” (CVE-2021-44228) (#39)
The first part of this fix is a port of two fixes from azure-cosmos-cassandra-driver-3-extensions. See PR #37:

- The position of the primary region is preserved in the preferred region list.

  CosmosLoadBalancingPolicy adds the primary region to the end of the preferred region list, if it's not specified. It now 
  preserves the primary region's position in the list when it is specified. Prior to this change, it incorrectly moved the 
  primary region to the end of the list.

- CosmosLoadBalancingPolicy::onDown now removes hosts based on endpoint address, not datacenter name.

  This behavior change avoids a problem that arises when a host is removed before its datacenter name has been
  determined. It also ensures that the address of the host removed matches the address of the host to be
  removed.

- Tests more thoroughly ensure that CosmosLoadBalancingPolicy orders hosts correctly based on the specification of
  preferred regions.

The second part of this fix addresses Log4j2 Vulnerability “Log4Shell” (CVE-2021-44228). This vulnerability was reported between late November and early December 2021. It is described in the National Vulnerabilities Database.
On 12/14/2021 Apache released version Apache log4j2 2.16.0 to completely remove support for Message Lookups and
disable JNDI by default. 

The test and example code in this repository depend on Apache log4j2 and this release bumps the version number for log4j2 from 2.13 to 2.16. The product code takes no dependency on log4j2. It uses slf4j-api instead.