Address Log4j2 Vulnerability “Log4Shell” (CVE-2021-44228) (#44)
The vulnerability that this change addresses is described in the [National Vulnerabilities Database][2]. On 12/14/2021
Apache released version [Apache log4j2][3] 2.16.0 to completely remove support for Message Lookups and disable JNDI by
default. The test and example code in this repository depend on Apache log4j2 and this release bumps the version
number for log4j2 from 2.13 to 2.16. The product code takes no dependency on log4j2. It uses [slf4j-api][4] instead.

[1]: https://nvd.nist.gov/vuln/detail/CVE-2021-44228
[2]: https://nvd.nist.gov/
[3]: https://github.com/apache/logging-log4j2
[4]: http://www.slf4j.org